+++ packages/gestalt/src/TextField/__tests__/TextField.test.js
@@ -3,21 +3,21 @@ import React from 'react';
 import { create } from 'react-test-renderer';
 import { shallow } from 'enzyme';
 import TextField from '../TextField';
-import ErrorFlyout from '../../ErrorFlyout/ErrorFlyout';
+import Flyout from '../../Flyout/Flyout';
 
 describe('TextField', () => {
-  it('Renders an ErrorFlyout if an error message is passed in', () => {
+  it('Renders a Flyout if an error message is passed in', () => {
     const wrapper = shallow(
       <TextField errorMessage="test" id="test" onChange={jest.fn()} />
     );
     wrapper.instance().setState({ errorIsOpen: true });
     wrapper.simulate('focus');
-    expect(wrapper.find(ErrorFlyout)).toHaveLength(1);
+    expect(wrapper.find(Flyout)).toHaveLength(1);
   });
 
-  it('Does not render an ErrorFlyout when errorMessage is null', () => {
+  it('Does not render a Flyout when errorMessage is null', () => {
     const wrapper = shallow(<TextField id="test" onChange={jest.fn()} />);
-    expect(wrapper.find(ErrorFlyout)).toHaveLength(0);
+    expect(wrapper.find(Flyout)).toHaveLength(0);
   });
 
   it('TextField normal', () => {
@@ -80,11 +80,11 @@ describe('TextField', () => {
         onBlur={jest.fn()}
       />
     );
-    expect(tree.find(ErrorFlyout)).toHaveLength(0);
+    expect(tree.find(Flyout)).toHaveLength(0);
     tree.setProps({
       errorMessage: 'error message',
     });
-    expect(tree.find(ErrorFlyout)).toHaveLength(1);
+    expect(tree.find(Flyout)).toHaveLength(1);
   });
 
   it('TextField with type number', () => {